remove some slag and move all front-end to client folder
