+++ interlecture/client/core/main.jsx
@@ -1,0 +1,31 @@
+import React from 'react';
+import { render } from 'react-dom';
+import { createStore } from 'redux';
+import { Provider } from 'react-redux';
+import Classroom from '../questions/classroom';
+import SocketHandler from './socketHandler';
+import questions from './reducers';
+
+const store = createStore(questions);
+const sock = new SocketHandler(`ws://${window.location.host}`, store);
+
+function onSubmit(event) {
+  sock.handleSubmit(event);
+}
+
+const classroom = {
+  id: 0,
+  title: 'test',
+  lecturer: 'gruppe69',
+};
+
+const Main = () => (
+  <Provider store={store}>
+    <div style={{ margin: '5pt 5%' }}>
+      <a href="/logout/">Logout</a>
+      <Classroom classroom={classroom} onSubmit={onSubmit} />
+
+    </div>
+  </Provider>
+);
+render(<Main />, document.getElementById('main'));
