--- interlecture/engine/js/fixtures.jsx
@@ -1,30 +1,0 @@
-const questions = [
-  {
-    id: 1,
-    user: 'Rolf',
-    body: 'hvorfor gjør vi dette?',
-  },
-  {
-    id: 2,
-    user: 'Arne',
-    body: 'kommer dette på eksamen?',
-  },
-  {
-    id: 3,
-    user: 'Kåre',
-    body: 'hva er meningen med livet?',
-  },
-  {
-    id: 4,
-    user: 'Bjarne',
-    body: 'finnes egentlig julenissen?',
-  },
-];
-
-const classroom = {
-  id: 0,
-  title: 'testklasserom',
-  lecturer: 'gruppe 69',
-};
-
-export { classroom, questions };
